requirements.txt
# Core dependencies
click>=8.1.0
PyYAML>=6.0
pyperclip>=1.8.2

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.4.0
mkdocs-minify-plugin>=0.7.0
mkdocs-git-revision-date-localized-plugin>=1.2.0
mkdocs-awesome-pages-plugin>=2.9.0

# Optional but recommended
pygments>=2.16.0
pymdown-extensions>=10.3.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.1.0
mypy>=1.5.0
